+++ src/commands/index.ts
@@ -18,6 +18,7 @@ export * from "./DeleteUnifiedCommand";
 export * from "./DuplicateCommand";
 export * from "./FocusCommand";
 export * from "./InstallWorkspaceTemplateFilesCommand";
+export * from "./InvalidateNugetCacheCommand";
 export * from "./MoveCommand";
 export * from "./MoveFileUpCommand";
 export * from "./MoveFileDownCommand";