+++ src/language/decorators/ICodeDecorator.ts
@@ -1,0 +1,6 @@
+import * as vscode from "vscode";
+
+export interface ICodeDecorator {
+    get filter(): vscode.DocumentSelector;
+    decorate(editor: vscode.TextEditor): void | Promise<void>;
+}