+++ src/language/filters.ts
@@ -1,0 +1,3 @@
+import * as vscode from "vscode";
+
+export const CSPROJ: vscode.DocumentSelector = { pattern: '**/*.csproj' };